Only build openssl if not cached

+++ .travis/install.sh
@@ -5,19 +5,27 @@ if [ -z "${OPENSSL_VERSION}" ]; then
   exit
 fi
 
+echo -en "travis_fold:start:openssl-install\\r"
+echo "Installing openssl $OPENSSL_VERSION"
+
 OPENSSL_DIR=$HOME/openssl/$OPENSSL_VERSION
 
-curl -O https://www.openssl.org/source/openssl-$OPENSSL_VERSION.tar.gz
-tar -zxf openssl-$OPENSSL_VERSION.tar.gz
-cd openssl-$OPENSSL_VERSION
-./config shared no-asm no-ssl2 no-ssl3 -fPIC --prefix="$OPENSSL_DIR"
+if [ ! -f "$OPENSSL_DIR/bin/openssl" ]
+then
+  curl -O https://www.openssl.org/source/openssl-$OPENSSL_VERSION.tar.gz
+  tar -zxf openssl-$OPENSSL_VERSION.tar.gz
+  cd openssl-$OPENSSL_VERSION
+  ./config shared no-asm no-ssl2 no-ssl3 -fPIC --prefix="$OPENSSL_DIR"
 
-# modify the shlib version to a unique one to make sure the dynamic
-# linker doesn't load the system one. This isn't required for 1.1.0 at the
-# moment since our Travis builders have a diff shlib version, but it doesn't hurt
-sed -i "s/^SHLIB_MAJOR=.*/SHLIB_MAJOR=100/" Makefile
-sed -i "s/^SHLIB_MINOR=.*/SHLIB_MINOR=0.0/" Makefile
-sed -i "s/^SHLIB_VERSION_NUMBER=.*/SHLIB_VERSION_NUMBER=100.0.0/" Makefile
+  # modify the shlib version to a unique one to make sure the dynamic
+  # linker doesn't load the system one. This isn't required for 1.1.0 at the
+  # moment since our Travis builders have a diff shlib version, but it doesn't hurt
+  sed -i "s/^SHLIB_MAJOR=.*/SHLIB_MAJOR=100/" Makefile
+  sed -i "s/^SHLIB_MINOR=.*/SHLIB_MINOR=0.0/" Makefile
+  sed -i "s/^SHLIB_VERSION_NUMBER=.*/SHLIB_VERSION_NUMBER=100.0.0/" Makefile
 
-make depend
-make install
+  make depend
+  make install
+fi
+
+echo -en "travis_fold:end:openssl-install\\r"
